Refactor STT feature code
