+++ frontend/src/components/editor/Editor.tsx
@@ -5,7 +5,7 @@ import { vim } from "@replit/codemirror-vim";
 import { basicSetup } from "@uiw/codemirror-extensions-basic-setup";
 import { xcodeDark, xcodeLight } from "@uiw/codemirror-theme-xcode";
 import { EditorView } from "codemirror";
-import { useCallback, useEffect, useState, useRef } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { ScrollSyncPane } from "react-scroll-sync";
 import { useCreateUploadUrlMutation, useUploadFileMutation } from "../../hooks/api/file";
@@ -23,35 +23,7 @@ import { yorkieCodeMirror } from "../../utils/yorkie";
 import EditorBottomBar, { BOTTOM_BAR_HEIGHT } from "./EditorBottomBar";
 import ToolBar from "./ToolBar";
 import SpeechToTextButton from "../common/SpeechToTextButton";
-
-interface SpeechRecognition extends EventTarget {
-	continuous: boolean;
-	interimResults: boolean;
-	lang: string;
-	start(): void;
-	stop(): void;
-	abort(): void;
-	onresult: (event: SpeechRecognitionEvent) => void;
-	onerror: (event: SpeechRecognitionErrorEvent) => void;
-	onend: () => void;
-}
-
-interface SpeechRecognitionEvent {
-	results: SpeechRecognitionResultList;
-	resultIndex: number;
-}
-
-interface SpeechRecognitionErrorEvent {
-	error: string;
-	message: string;
-}
-
-declare global {
-	interface Window {
-		SpeechRecognition: new () => SpeechRecognition;
-		webkitSpeechRecognition: new () => SpeechRecognition;
-	}
-}
+import { useSpeechToText } from "./SpeechToTextFeature";
 
 interface EditorProps {
 	width: number | string;
@@ -70,103 +42,17 @@ function Editor(props: EditorProps) {
 	const { mutateAsync: uploadFile } = useUploadFileMutation();
 	const { applyFormat, setKeymapConfig } = useFormatUtils();
 	const { toolBarState, setToolBarState, updateFormatBar } = useToolBar();
-	const [isListening, setIsListening] = useState(false);
-	const recognitionRef = useRef<SpeechRecognition | null>(null);
-	const [interimTranscript, setInterimTranscript] = useState("");
-	const finalTranscriptRef = useRef("");
+
+	const {
+		state: speechToTextState,
+		toggleListening,
+		isSpeechRecognitionSupported,
+	} = useSpeechToText();
 
 	const ref = useCallback((node: HTMLElement | null) => {
 		if (!node) return;
 		setElement(node);
 	}, []);
-
-	useEffect(() => {
-		const SpeechRecognition = window.SpeechRecognition || window.webkitSpeechRecognition;
-		if (SpeechRecognition) {
-			recognitionRef.current = new SpeechRecognition();
-			recognitionRef.current.continuous = true;
-			recognitionRef.current.interimResults = true;
-			recognitionRef.current.lang = "ko-KR";
-		}
-
-		return () => {
-			if (recognitionRef.current) {
-				recognitionRef.current.abort();
-			}
-		};
-	}, []);
-
-	useEffect(() => {
-		if (!recognitionRef.current || !editorStore.cmView || !editorStore.doc) return;
-
-		recognitionRef.current.onresult = (event) => {
-			let interimTranscript = "";
-			let finalTranscript = "";
-
-			for (let i = event.resultIndex; i < event.results.length; i++) {
-				const transcript = event.results[i][0].transcript;
-				if (event.results[i].isFinal) {
-					finalTranscript += transcript + " ";
-				} else {
-					interimTranscript += transcript;
-				}
-			}
-
-			if (finalTranscript) {
-				finalTranscriptRef.current = finalTranscript;
-
-				const cursor = editorStore.cmView!.state.selection.main;
-				const from = cursor.from;
-				const to = cursor.to;
-				const newCursorPos = from + finalTranscript.length;
-
-				editorStore.doc?.update((root, presence) => {
-					root.content.edit(from, to, finalTranscript);
-					presence.set({
-						selection: root.content.indexRangeToPosRange([newCursorPos, newCursorPos]),
-					});
-				});
-
-				editorStore.cmView!.dispatch({
-					changes: { from, to, insert: finalTranscript },
-					selection: { anchor: newCursorPos, head: newCursorPos },
-				});
-			}
-
-			setInterimTranscript(interimTranscript);
-		};
-
-		recognitionRef.current.onerror = (event) => {
-			console.error("Speech recognition error:", event.error);
-			setIsListening(false);
-		};
-
-		recognitionRef.current.onend = () => {
-			setIsListening(false);
-			setInterimTranscript("");
-			finalTranscriptRef.current = "";
-		};
-	}, [editorStore.cmView, editorStore.doc]);
-
-	// Toggle speech recognition
-	const toggleListening = useCallback(() => {
-		if (!recognitionRef.current) {
-			// Show browser compatibility warning if Speech Recognition is not available
-			alert(
-				"Speech recognition is not supported in your browser. Please try using Chrome, Edge, or Safari."
-			);
-			return;
-		}
-
-		if (isListening) {
-			recognitionRef.current.stop();
-			setInterimTranscript("");
-		} else {
-			recognitionRef.current.start();
-			setIsListening(true);
-			finalTranscriptRef.current = "";
-		}
-	}, [isListening]);
 
 	useEffect(() => {
 		if (
@@ -263,11 +149,13 @@ function Editor(props: EditorProps) {
 								onChangeToolBarState={setToolBarState}
 							/>
 						)}
-						<SpeechToTextButton
-							isListening={isListening}
-							onClick={toggleListening}
-							interimTranscript={interimTranscript}
-						/>
+						{isSpeechRecognitionSupported && (
+							<SpeechToTextButton
+								isListening={speechToTextState.isListening}
+								onClick={toggleListening}
+								interimTranscript={speechToTextState.interimTranscript}
+							/>
+						)}
 					</div>
 				</ScrollSyncPane>
 			</div>
